Refactoring
Remove GetSimpleCoveringId and FloodFillId, various S2 improvements, add
custom array.array headers with 64-bit integer support,
expand __init__.py, remove pogeo.pyx and move its parts into multiple
files, expand S2 pxds.

+++ geometry/s2cellunion.cc
@@ -23,7 +23,7 @@ using std::vector;
 // Returns true if the vector of cell_ids is sorted.  Used only in
 // DCHECKs.
 extern bool IsSorted(vector<S2CellId> const& cell_ids) {
-  for (int i = 0; i + 1 < cell_ids.size(); ++i) {
+  for (size_t i = 0; i + 1 < cell_ids.size(); ++i) {
     if (cell_ids[i + 1] < cell_ids[i]) {
       return false;
     }
